requirements.txt
# Dépendances principales
aiohttp==3.8.6
beautifulsoup4==4.12.2
openai==1.3.8
selenium==4.15.2

# Dépendances utilitaires
python-dotenv==1.0.0
requests==2.31.0
lxml==4.9.3

# Logging et gestion des erreurs
loguru==0.7.2

# Développement et tests
pytest==7.4.3
pytest-asyncio==0.21.1 